UserPreference split into provider/controller
