+++ lib/Alchemy/Phrasea/Controller/User/UserPreferenceController.php
@@ -1,0 +1,72 @@
+<?php
+/*
+ * This file is part of Phraseanet
+ *
+ * (c) 2005-2015 Alchemy
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+namespace Alchemy\Phrasea\Controller\User;
+
+use Alchemy\Phrasea\Application;
+use Alchemy\Phrasea\Controller\Controller;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+
+class UserPreferenceController extends Controller
+{
+    /**
+     *  Save temporary user preferences
+     *
+     * @param  Application  $app
+     * @param  Request      $request
+     * @return JsonResponse
+     */
+    public function saveTemporaryPref(Application $app, Request $request)
+    {
+        if (!$request->isXmlHttpRequest()) {
+            $app->abort(400);
+        }
+
+        $prop = $request->request->get('prop');
+        $value = $request->request->get('value');
+        $success = false;
+        $msg = $app->trans('Error while saving preference');
+
+        if ($prop && $value) {
+            $app['session']->set('phraseanet.' . $prop, $value);
+            $success = true;
+            $msg = $app->trans('Preference saved !');
+        }
+
+        return new JsonResponse(['success' => $success, 'message' => $msg]);
+    }
+
+    /**
+     *  Save user preferenes
+     *
+     * @param  Application  $app
+     * @param  Request      $request
+     * @return JsonResponse
+     */
+    public function saveUserPref(Application $app, Request $request)
+    {
+        if (!$request->isXmlHttpRequest()) {
+            $app->abort(400);
+        }
+
+        $msg = $app->trans('Error while saving preference');
+        $prop = $request->request->get('prop');
+        $value = $request->request->get('value');
+
+        $success = false;
+        if (null !== $prop && null !== $value) {
+            $app['manipulator.user']->setUserSetting($app['authentication']->getUser(), $prop, $value);
+            $success = true;
+            $msg = $app->trans('Preference saved !');
+        }
+
+        return new JsonResponse(['success' => $success, 'message' => $msg]);
+    }
+}
